Remove chunk size assumptions

+++ src/test/java/com/voxelplugineering/voxelsniper/MemoryContainerTest.java
@@ -1,3 +1,26 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2014 The Voxel Plugineering Team
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 package com.voxelplugineering.voxelsniper;
 
 import static org.junit.Assert.assertEquals;
@@ -28,7 +51,7 @@ public class MemoryContainerTest
         container.writeBoolean("a", true);
         container.writeByte("b", (byte) 0);
         container.writeShort("c", (short) 0);
-        container.writeInt("d", (int) 0);
+        container.writeInt("d", 0);
         container.writeLong("e", (long) 0);
         container.writeFloat("f", (float) 0);
         container.writeDouble("g", (double) 0);
@@ -40,7 +63,7 @@ public class MemoryContainerTest
         assertEquals(Boolean.valueOf(true), container.readBoolean("a").get());
         assertEquals(Byte.valueOf((byte) 0), container.readByte("b").get());
         assertEquals(Short.valueOf((short) 0), container.readShort("c").get());
-        assertEquals(Integer.valueOf((int) 0), container.readInt("d").get());
+        assertEquals(Integer.valueOf(0), container.readInt("d").get());
         assertEquals(Long.valueOf((long) 0), container.readLong("e").get());
         assertEquals(Float.valueOf((float) 0), container.readFloat("f").get());
         assertEquals(Double.valueOf((double) 0), container.readDouble("g").get());
@@ -62,7 +85,7 @@ public class MemoryContainerTest
         container.writeBoolean("a", true);
         container.writeByte("b", (byte) 0);
         container.writeShort("c", (short) 0);
-        container.writeInt("d", (int) 0);
+        container.writeInt("d", 0);
         container.writeLong("e", (long) 0);
         container.writeFloat("f", (float) 0);
         container.writeDouble("g", (double) 0);
@@ -82,7 +105,7 @@ public class MemoryContainerTest
         assertEquals(Boolean.valueOf(true), newContainer.readBoolean("a").get());
         assertEquals(Byte.valueOf((byte) 0), newContainer.readByte("b").get());
         assertEquals(Short.valueOf((short) 0), newContainer.readShort("c").get());
-        assertEquals(Integer.valueOf((int) 0), newContainer.readInt("d").get());
+        assertEquals(Integer.valueOf(0), newContainer.readInt("d").get());
         assertEquals(Long.valueOf((long) 0), newContainer.readLong("e").get());
         assertEquals(Float.valueOf((float) 0), newContainer.readFloat("f").get());
         assertEquals(Double.valueOf((double) 0), newContainer.readDouble("g").get());
